refactor(exception): extract global exception handling to dedicated advice package feat(contact): add EmailSendFailedException and improve email error handling
- Moved GlobalExceptionHandler, ApiErrorResponse, and ValidationError into new
  top-level `advice` package for true global scope and clearer architecture
- Added domain-specific EmailSendFailedException under contact/exception
- Updated EmailService to throw EmailSendFailedException on Resend API failures
  to ensure proper error handling and backend->frontend error communication
- Improved backend consistency by separating global error handling from
  feature-specific exceptions
