+++ src/main/java/com/david/examportfolio/exam_portfolio_backend/contact/exception/EmailSendFailException.java
@@ -1,0 +1,12 @@
+package com.david.examportfolio.exam_portfolio_backend.contact.exception;
+
+public class EmailSendFailException extends RuntimeException {
+
+    public EmailSendFailException(String message) {
+        super(message);
+    }
+
+    public EmailSendFailException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
